+++ colisoes.c
@@ -6,144 +6,7 @@
 #define LARG 10
 #define ALT 20
 
-//imprime o menu
-void imprimirMenu(long score, long highscore, int nivel){
-  clear();
-  //menu
-  move(2,2);
-  printw("TETRIS");
-  move(3,2);
-  printw("Feito por:");
-  move(4,2);
-  printw("Antonio");
-  move(5,2);
-  printw("Ademario Junior");
-  move(6,2);
-  printw("Danilo Felipe");
-  move(7,2);
-  printw("Francisco Guilherme");
-  move(8,2);
-  printw("Joao Macedo");
-  move(17,49);
-  printw("Para parar o jogo aperte P");
-  move(19,49);
-  printw("Para sair do jogo aperte Q");
-  move(2,55);
-  printw("Nivel %d", nivel);
-  move(3,55);
-  printw("SCORE");
-  move(4,55);
-  printw("%ld", score);
-  move(6,55);
-  printw("HIGHSCORE");
-  move(7, 55);
-  if (score > highscore)  printw("%ld", score);
-  else printw("%ld", highscore);
 
-  refresh();
-}
-
-//imprime a peca e o tabuleiro
-void imprimirJogo(WINDOW * win, char tabuleiro[LARG][ALT], char peca[4][4], int eixoX, int eixoY){
-  int x, y;
-  //imprime a matriz na janela win
-  for (y = 0; y < ALT ; y++) {
-    for (x = 0; x < LARG; x++) {
-      //move o cursor na janela win
-      wmove(win, y, x);
-      wprintw(win, "%c", tabuleiro[x][y]);
-    }
-  }
-
-  //imprime a peca na janela win
-  for (y = eixoY; y < eixoY + 4; y++) {
-    for (x = eixoX; x < eixoX + 4; x++) {
-      wmove(win, y, x);
-      wprintw(win, "%c", peca[x - eixoX][y - eixoY]);
-    }
-  }
-
-  wrefresh(win);
-}
-
-//recebe o highscore
-int recebeHighscore(FILE * myFile){
-  long highscore;
-  //abre o arquivo
-  myFile = fopen("rank.txt", "r");
-
-  //pega o valor que esta no arquivo e coloca na variavel highscore
-  fscanf(myFile, "%ld", &highscore);    
-
-  //fecha o arquivo
-  fclose(myFile);
-
-  return highscore;
-}
-
-
-//salva o highscore
-void salvarHighscore(FILE * myFile, long score){
-  long highscore;
-
-  myFile = fopen("rank.txt", "r");
-
-  fscanf(myFile, "%ld", &highscore);    
-
-  fclose(myFile);
-
-  if(score > highscore){
-    myFile = fopen("rank.txt", "w");
-
-    fprintf(myFile, "%ld", score);
-
-    fclose(myFile);
-  }
-}
-
-//modifica a velocidade com que a pedra cai
-int modificaVelocidade(int contador){
-  if (contador < 5)       return 1000;
-  else if(contador < 15)  return 800;
-  else if(contador < 25)  return 700;
-  else if(contador < 35)  return 600;
-  else if(contador < 45)  return 500;
-  else if(contador < 55)  return 400;
-  else if(contador < 65)  return 325;
-  else if(contador < 75)  return 250;
-  else if(contador < 85)  return 200;
-  else if(contador < 95)  return 150;
-  else                    return 100;
-}
-
-//modifica o nivel
-int modificaNivel(int contador){
-  if (contador < 5)       return 1;
-  else if(contador < 15)  return 2;
-  else if(contador < 25)  return 3;
-  else if(contador < 35)  return 4;
-  else if(contador < 45)  return 5;
-  else if(contador < 55)  return 6;
-  else if(contador < 65)  return 7;
-  else if(contador < 75)  return 8;
-  else if(contador < 85)  return 9;
-  else if(contador < 95)  return 10;
-  else                    return 11;
-}
-
-//verifica se a peca pode ser criada
-int finalizar(char peca [4][4], char tabuleiro[LARG][ALT], int eixoX, int eixoY){
-  char y, x;
-  for (y = 0; y < 4; ++y){
-    for (x = 0; x < 4; ++x){
-      if(tabuleiro[eixoX + x][eixoY + y] == PEDRA && peca[x][y] == PEDRA){
-        return 1;
-      }
-    }
-  }
-
-  return 0;
-}
 
 //impede a peca de sair pela direita do tabuleiro
 int verificaD(char peca[4][4]){
