+++ colisoes.h
@@ -6,21 +6,6 @@
 #define LARG 10
 #define ALT 20
 
-//imprime o menu
-void imprimirMenu(long score, long highscore, int nivel);
-//imprime o jogo
-void imprimirJogo(WINDOW *win, char tabuleiro[LARG][ALT], char peca[4][4], int eixoX, int eixoY);
-//recebe o highscore
-int recebeHighscore(FILE * myFile);
-//salva o highscore
-void salvarHighscore(FILE * myFile, long score);
-//modifica a velocidade com que a pedra cai
-int modificaVelocidade(int contador);
-//modifica o nivel do jogo
-int modificaNivel(int contador);
-
-//verifica se a pedra pode ser criada
-int finalizar(char peca [4][4], char tabuleiro[LARG][ALT], int eixoX, int eixoY);
 
 //impede a peca de sair pela direita do tabuleiro
 int verificaD(char peca[4][4]);
